Remove extraneous *Impl classes, controller factory, merge medSql into medCore
The Impl classes were just adding bloat for no good reason, and the factory was used by nothing but the data manager, so might as well get rid of it. This required medSql to move into medCore, as now the data manager needed to access the ::instance() of each controller directly, requiring a link on medSql from medCore, so might as well merge.